breaking fix so gaomon s56k works
As per https://github.com/DIGImend/digimend-kernel-drivers/issues/305,
there are issues with using the GAOMON S56K tablet. This commit fixes
this by deleting all the stuff related to probing v2 pen parameters,
which seems to be the issue.

I don't know how to actually fix this but this works for me on kernel
`5.6.15-arch1-1`.